+++ elysia/controllers/communities.ts
@@ -1,0 +1,3 @@
+export const communities = (await fetch(
+  "https://discuit.net/api/communities",
+).then((r) => r.json())) as Community[];
