+++ elysia/main.tsx
@@ -1,59 +1,14 @@
 import { Elysia } from "elysia";
 import { html } from "@elysiajs/html";
 import { staticPlugin } from "@elysiajs/static";
-import Index from "./routes/index";
-import { communities, search } from "./controllers/search";
-import Community from "./components/Community";
+import index from "./routes/index";
+import searchRoute from "./routes/search";
 
 const app = new Elysia()
   .use(html())
   .use(staticPlugin())
-  .get("/", () => <Index />)
-  .get("/search", async ({ query }) => {
-    if (query["q"] === "" || !query["q"]) {
-      return (
-        <>
-          {communities.map(({ name, about, noMembers, proPic }) => (
-            <Community
-              name={name}
-              about={about}
-              noMembers={noMembers}
-              proPic={proPic}
-            />
-          ))}
-        </>
-      );
-    }
-
-    const searchResults = await search(query["q"]);
-    if (!searchResults) {
-      return (
-        <>
-          {communities.map(({ name, about, noMembers, proPic }) => (
-            <Community
-              name={name}
-              about={about}
-              noMembers={noMembers}
-              proPic={proPic}
-            />
-          ))}
-        </>
-      );
-    }
-
-    return (
-      <>
-        {searchResults.map(({ name, about, noMembers, proPic }) => (
-          <Community
-            name={name}
-            about={about}
-            noMembers={noMembers}
-            proPic={proPic}
-          />
-        ))}
-      </>
-    );
-  })
+  .use(index)
+  .use(searchRoute)
   .listen(8080);
 
 console.log(`Listening on ${app.server!.url}`);
