+++ elysia/package.json
@@ -14,6 +14,7 @@
     "@elysiajs/static": "^1.0.2",
     "@kitajs/ts-html-plugin": "^4.0.1",
     "elysia": "^1.0.13",
+    "lucide-static": "^0.366.0",
     "marked": "^12.0.1",
     "meilisearch": "^0.38.0"
   }
