+++ elysia/routes/index.tsx
@@ -1,7 +1,9 @@
+import { Elysia } from "elysia";
 import Community from "../components/Community";
-import { communities } from "../controllers/search";
+import { communities } from "../controllers/communities";
+import { Sun, Moon, Github } from "lucide-static";
 
-export default function () {
+export default new Elysia().get("/", () => {
   return (
     <html lang="en" class={"hidden"} _={`on load remove .hidden from me`}>
       <head>
@@ -19,10 +21,11 @@ export default function () {
       <body>
         <header>
           <h1>Discuit Search</h1>
-          <div class={"buttons"}>
+          <div>
             <button
               type="button"
               class={"theme-toggle"}
+              title="Theme toggle"
               _={`
               on load
                 if matchMedia("(prefers-color-scheme: dark)").matches
@@ -32,53 +35,14 @@ export default function () {
               on click toggle between .dark and .light on <html />
               `}
             >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                stroke-width="2"
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                class="lucide lucide-sun"
-              >
-                <circle cx="12" cy="12" r="4" />
-                <path d="M12 2v2" />
-                <path d="M12 20v2" />
-                <path d="m4.93 4.93 1.41 1.41" />
-                <path d="m17.66 17.66 1.41 1.41" />
-                <path d="M2 12h2" />
-                <path d="M20 12h2" />
-                <path d="m6.34 17.66-1.41 1.41" />
-                <path d="m19.07 4.93-1.41 1.41" />
-              </svg>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                stroke-width="2"
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                class="lucide lucide-moon"
-              >
-                <path d="M12 3a6 6 0 0 0 9 9 9 9 0 1 1-9-9Z" />
-              </svg>
+              {Sun}
+              {Moon}
             </button>
-            <a href="https://github.com/Codycody31/discuit-search">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                stroke-width="2"
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                class="lucide lucide-github"
-              >
-                <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4" />
-                <path d="M9 18c-4.51 2-5-2-7-2" />
-              </svg>
+            <a
+              href="https://github.com/Codycody31/discuit-search"
+              title="GitHub link"
+            >
+              {Github}
             </a>
           </div>
         </header>
@@ -91,19 +55,21 @@ export default function () {
             hx-trigger="keyup changed"
             hx-include="this"
             placeholder="Search..."
+            title="Search for a community"
           />
-          <ul id="results">
-            {communities.map(({ name, about, noMembers, proPic }) => (
+          <div id="results">
+            {communities.map(({ id, name, about, noMembers, proPic }) => (
               <Community
+                id={id}
                 name={name}
                 about={about}
                 noMembers={noMembers}
                 proPic={proPic}
               />
             ))}
-          </ul>
+          </div>
         </main>
       </body>
     </html>
   );
-}
+});
