+++ elysia/types.d.ts
@@ -1,0 +1,9 @@
+type Community = {
+  id: string;
+  name: string;
+  about: string | null;
+  noMembers: number;
+  proPic: {
+    copies: [{ name: "medium"; url: string }];
+  } | null;
+};
